UPDATE: cache clear bug resolved
When a service has a version number then its cache is being saved as v1.service.find instead of service.find
this version number of the service is present in fullName if a version number is used, otherwise fullName only contain name of service. So using fullName instead of name in clearCache resolves this problem.