Update README.md
Currently specified mongoose version is not compiling with typescript because of an old 'bson' package. on 5.13.9 release this dependency has been updated and hence, fixed the problem. Currently the latest 5.13.x release is 5.13.14 also tagged as 'legacy'. Spent hours experimenting on which version to use with moleculer-db-adapter-mongoose, so that information could be useful for me and others. Don't see any blocking problems to updating the version